feat(log): use LogBuilder to initialize logging
