+++ sn_logging/Cargo.toml
@@ -11,6 +11,8 @@ repository = "https://github.com/maidsafe/safe_network"
 version = "0.2.12"
 
 [dependencies]
+chrono = "~0.4.19"
+dirs-next = "~2.0.0"
 file-rotate = "0.7.3"
 opentelemetry = { version = "0.20", features = ["rt-tokio"], optional = true }
 opentelemetry-otlp = { version = "0.13", optional = true }
