+++ sn_node/tests/sequential_transfers.rs
@@ -9,14 +9,15 @@
 mod common;
 
 use assert_fs::TempDir;
-use common::{get_client_and_wallet, get_wallet, init_logging_single_threaded_tokio};
+use common::{get_client_and_wallet, get_wallet};
 use eyre::Result;
 use sn_client::send;
+use sn_logging::LogBuilder;
 use sn_transfers::{create_offline_transfer, rng, Hash, NanoTokens, UniquePubkey};
 
 #[tokio::test]
 async fn cash_note_transfer_multiple_sequential_succeed() -> Result<()> {
-    let _log_guards = init_logging_single_threaded_tokio("sequential_transfer");
+    let _log_guards = LogBuilder::init_single_threaded_tokio_test("sequential_transfer");
 
     let first_wallet_balance = 1_000_000_000;
     let first_wallet_dir = TempDir::new()?;
@@ -53,7 +54,7 @@ async fn cash_note_transfer_multiple_sequential_succeed() -> Result<()> {
 
 #[tokio::test]
 async fn cash_note_transfer_double_spend_fail() -> Result<()> {
-    let _log_guards = init_logging_single_threaded_tokio("sequential_transfer");
+    let _log_guards = LogBuilder::init_single_threaded_tokio_test("sequential_transfer");
 
     // create 1 wallet add money from faucet
     let first_wallet_balance = 1_000_000_000;
